2020_115

+++ WPT_sotu_zyu/WPT_sotu_zyu.ino
@@ -1,7 +1,7 @@
 #include<MsTimer2.h>
 float f = 0;
 String data;
-int data0 = 0;
+float data0 = 0;
 void setup() {
   Serial.begin(57600);
   MsTimer2::set(100,flash);
@@ -29,5 +29,5 @@ void loop() {
 void receive_duty_data() {
   
   data = Serial.readString();
-  data0 = data.toInt();
+  data0 = data.toFloat();
 }
